requirements.txt
python-twitter
python-decouple
discord.py